CRM-17: frontend: delete process

+++ app/Livewire/Admin/Users/Delete.php
@@ -3,17 +3,21 @@
 namespace App\Livewire\Admin\Users;
 
 use App\Models\User;
-use Livewire\Component;
-use Livewire\Attributes\Validate;
-use Illuminate\Contracts\View\View;
 use App\Notifications\UserDeletedNotification;
+use Illuminate\Contracts\View\View;
+use Livewire\Attributes\{On, Rule};
+use Livewire\Component;
+use Mary\Traits\Toast;
 
 class Delete extends Component
 {
-    public User $user;
+    use Toast;
+
+    public ?User $user = null;
+
     public bool $modal = false;
 
-    #[Validate(['required', 'confirmed'])]
+    #[Rule(['required', 'confirmed'])]
     public string $confirmation = 'DART VADER';
 
     public ?string $confirmation_confirmation = null;
@@ -23,12 +27,29 @@ class Delete extends Component
         return view('livewire.admin.users.delete');
     }
 
+    #[On('user::deletion')]
+    public function openConfirmationFor(int $userId): void
+    {
+        $this->user  = User::select('id', 'name')->find($userId);
+        $this->modal = true;
+    }
+
     public function destroy(): void
     {
         $this->validate();
-        $this->user->delete();
-        $this->user->notify(new UserDeletedNotification());
 
+        if($this->user->is(auth()->user())) {
+
+            $this->addError('confirmation', "You can't delete yourself brow.");
+
+            return;
+        }
+
+        $this->user->delete();
+
+        $this->user->notify(new UserDeletedNotification());
+        $this->success('User deleted successfully.');
         $this->dispatch('user::deleted');
+        $this->reset('modal');
     }
 }
